+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Changelog for NeoFS Node
 - Synchronous objects `PUT` to the local storage (#2607)
 - `copies_number` field was not used in `PUT` request handling (#2607)
 - Neo-go notary deposit data workaround (#2429)
+- Error of missing network endpoint in `neofs-cli object` commands when it is set in the file config (#2608)
 
 ### Changed
 - FSTree storage now uses more efficient and safe temporary files under Linux (#2566)